Conflicts:
	src/console_gui.cpp
	src/genworld_gui.cpp
	src/group_gui.cpp
	src/misc_gui.cpp
	src/settings_gui.cpp
	src/widgets/settings_widget.h

+++ src/saveload/linkgraph_sl.cpp
@@ -16,12 +16,14 @@
 #include "../settings_internal.h"
 #include "saveload.h"
 
+#include "../safeguards.h"
+
 typedef LinkGraph::BaseNode Node;
 typedef LinkGraph::BaseEdge Edge;
 
 const SettingDesc *GetSettingDescription(uint index);
 
-static uint _num_nodes;
+static uint16 _num_nodes;
 
 /**
  * Get a SaveLoad array for a link graph.
